requirements.txt
hug==2.4.0
gunicorn==19.7.1
SQLAlchemy==1.2.6
psycopg2-binary==2.7.4
ujson==1.35
pendulum==1.4.4
python-dotenv==0.8.2
PyJWT==1.6.1
argon2-cffi==18.1.0
passlib==1.7.1